Print branch distance

+++ util/log.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
-int tracelog(int stmtid, int result, int branchdistance) {
+int tracelog(int stmtid, int result, double branchdistance) {
     FILE *f = fopen("trace", "a");
-    fprintf(f, "%d %d %d\n", stmtid, result, branchdistance);
+    fprintf(f, "%d %d %f\n", stmtid, result, branchdistance);
     fclose(f);
 }
